RavenDB-27169 Move UserDomainsAndLicenseInfo to its own file, rvn links SetupInfo but not LicenseStatus
