+++ src/Raven.Server/Commercial/SetupInfo.cs
@@ -320,21 +320,6 @@ namespace Raven.Server.Commercial
         }
     }
 
-    public sealed class UserDomainsAndLicenseInfo
-    {
-        public UserDomainsWithIps UserDomainsWithIps { get; set; }
-        public LicenseStatus LicenseStatus { get; set; }
-
-        public DynamicJsonValue ToJson()
-        {
-            return new DynamicJsonValue
-            {
-                [nameof(UserDomainsWithIps)] = UserDomainsWithIps.ToJson(),
-                [nameof(LicenseStatus)] = LicenseStatus?.ToJson()
-            };
-        }
-    }
-
     public sealed class UserDomainsWithIps
     {
         public string[] Emails { get; set; }
